requirements.txt
streamlit
moviepy
pandas
numpy
pillow
